+++ pkg/http/handler_test.go
@@ -40,6 +40,11 @@ func getTestAdmissionReviewV1beta1RequestStr(uid string) string {
 				Kind:    "Pod",
 				Version: "v1",
 			},
+			RequestKind: &metav1.GroupVersionKind{
+				Group:   "core",
+				Kind:    "Pod",
+				Version: "v1",
+			},
 			UID: types.UID(uid),
 			Object: runtime.RawExtension{
 				Object: &corev1.Pod{
@@ -64,6 +69,11 @@ func getTestAdmissionReviewV1RequestStr(uid string) string {
 		},
 		Request: &admissionv1.AdmissionRequest{
 			Kind: metav1.GroupVersionKind{
+				Group:   "core",
+				Kind:    "Pod",
+				Version: "v1",
+			},
+			RequestKind: &metav1.GroupVersionKind{
 				Group:   "core",
 				Kind:    "Pod",
 				Version: "v1",
@@ -248,8 +258,10 @@ func TestDefaultWebhookFlow(t *testing.T) {
 			// Mocks.
 			mwh := &webhookmock.Webhook{}
 			test.mock(mwh)
+			mwh.On("ID").Maybe().Return("")
+			mwh.On("Kind").Maybe().Return(model.WebhookKind(""))
 
-			h, err := kubewebhookhttp.HandlerFor(mwh)
+			h, err := kubewebhookhttp.HandlerFor(kubewebhookhttp.HandlerConfig{Webhook: mwh})
 			require.NoError(err)
 
 			req := httptest.NewRequest("GET", "/awesome/webhook", bytes.NewBufferString(test.body))